feat: finalize backend services and controllers

+++ server/src/api/compartment/compartment.service.ts
@@ -1,27 +1,46 @@
-import { BadRequestException, Injectable } from "@nestjs/common";
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from "@nestjs/common";
 import { PrismaService } from "src/infrastructure/prisma/prisma.service";
-import { CompartmentDto } from "./dto/compartment.dto";
+import { CompartmentDto, UpdateCompartmentDto } from "./dto/compartment.dto";
 
 @Injectable()
 export class CompartmentService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async getCompartments(unitId: number) {
-    return this.prisma.compartment.findMany({
+  async createCompartment(compartment: CompartmentDto) {
+    const { capacity, unitId } = compartment;
+
+    const unit = await this.prisma.unit.findUnique({
       where: {
-        unitId,
+        id: unitId,
+      },
+      include: {
+        compartments: {
+          orderBy: {
+            id: "asc",
+          },
+        },
       },
     });
-  }
 
-  async createUnit(compartment: CompartmentDto) {
-    const { capacity, macAddress, unitId } = compartment;
+    if (!unit) {
+      throw new BadRequestException("Unit does not exist.");
+    }
+
+    const count = unit.compartments.length;
+
+    if (count === 4) {
+      throw new BadRequestException("Unit is full.");
+    }
 
     try {
       return await this.prisma.compartment.create({
         data: {
           capacity,
-          macAddress,
+          macAddress: `${unit.macAddress}-${count + 1}`,
           unit: {
             connect: {
               id: unitId,
@@ -34,8 +53,8 @@ export class CompartmentService {
     }
   }
 
-  async updateCompartment(id: number, compartment: CompartmentDto) {
-    const { capacity, macAddress } = compartment;
+  async updateCompartment(id: number, compartment: UpdateCompartmentDto) {
+    const { capacity } = compartment;
 
     try {
       return await this.prisma.compartment.update({
@@ -44,7 +63,6 @@ export class CompartmentService {
         },
         data: {
           capacity,
-          macAddress,
         },
       });
     } catch (_e) {
@@ -53,10 +71,60 @@ export class CompartmentService {
   }
 
   async deleteCompartment(id: number) {
+    const compartment = await this.prisma.compartment.findUnique({
+      where: {
+        id,
+      },
+      select: {
+        unit: true,
+      },
+    });
+
+    if (!compartment) {
+      throw new NotFoundException("Compartment does not exist.");
+    }
+
+    const compartmentsInUnit = await this.prisma.compartment.findMany({
+      where: {
+        unitId: compartment.unit.id,
+      },
+      orderBy: {
+        id: "asc",
+      },
+    });
+
+    const remainingCompartments = compartmentsInUnit.filter(
+      (compartment) => compartment.id !== id,
+    );
+
     try {
-      return await this.prisma.compartment.delete({
+      await this.prisma.compartment.delete({
         where: {
           id,
+        },
+      });
+      return await this.prisma.unit.update({
+        where: {
+          id: compartment.unit.id,
+        },
+        data: {
+          compartments: {
+            update: remainingCompartments.map((c, index) => ({
+              where: {
+                id: c.id,
+              },
+              data: {
+                macAddress: `${compartment.unit.macAddress}-${index + 1}`,
+              },
+            })),
+          },
+        },
+        include: {
+          compartments: {
+            orderBy: {
+              id: "asc",
+            },
+          },
         },
       });
     } catch (_e) {
